requirements.txt
service_identity
scrapy
readability-lxml